First draft of test results view refactor
