+++ frontend/src/app/ReportingApp.tsx
@@ -16,8 +16,8 @@ import ProtectedRoute from "./commonComponents/ProtectedRoute";
 import Header from "./commonComponents/Header";
 import Page from "./commonComponents/Page/Page";
 import { setInitialState } from "./store";
-import TestResultsList from "./testResults/TestResultsList";
-import CleanTestResultsList from "./testResults/CleanTestResultsList";
+import TestResultsList from "./testResults/viewResults/TestResultsList";
+import CleanTestResultsList from "./testResults/viewResults/CleanTestResultsList";
 import TestQueueContainer from "./testQueue/TestQueueContainer";
 import ManagePatientsContainer from "./patients/ManagePatientsContainer";
 import EditPatientContainer from "./patients/EditPatientContainer";
