+++ frontend/src/app/testResults/mocks/queries.mock.tsx
@@ -9,7 +9,7 @@ import {
   GetResultsMultiplexWithCountDocument,
   ArchivedStatus,
 } from "../../../generated/graphql";
-import { testResultDetailsQuery } from "../TestResultDetailsModal";
+import { testResultDetailsQuery } from "../viewResults/actionMenuModals/TestResultDetailsModal";
 import { QUERY_PATIENT } from "../../testQueue/addToQueue/AddToQueueSearch";
 
 import testResults from "./resultsCovid.mock";
