+++ frontend/src/app/testResults/viewResults/TestResultsList.scss
@@ -1,4 +1,4 @@
-@use "../../scss/settings";
+@use "../../../scss/settings";
 
 .sr-test-results-list {
   .sticky-heading {
