+++ frontend/src/app/testResults/viewResults/actionMenuModals/TestResultPrintModal.scss
